Moved out MapEnums, moved TeleportToStartingPoint to Outdoor.h/cpp
