+++ src/Engine/Events/EventInterpreter.cpp
@@ -17,6 +17,7 @@
 #include "Engine/Tables/ItemTable.h"
 #include "Engine/Spells/Spells.h"
 #include "Engine/Engine.h"
+#include "Engine/MapInfo.h"
 
 #include "Media/Audio/AudioPlayer.h"
 #include "Media/MediaPlayer.h"
