+++ src/Engine/Graphics/Outdoor.cpp
@@ -35,6 +35,7 @@
 #include "Engine/TurnEngine/TurnEngine.h"
 #include "Engine/Graphics/Vis.h"
 #include "Engine/Graphics/BspRenderer.h"
+#include "Engine/MapInfo.h"
 
 #include "GUI/GUIProgressBar.h"
 #include "GUI/GUIWindow.h"
@@ -3006,4 +3007,47 @@ double OutdoorLocation::GetPolygonMaxZ(RenderVertexSoft *pVertex, unsigned int u
             result = pVertex[i].vWorldPosition.z;
     }
     return result;
+}
+
+void TeleportToStartingPoint(MapStartPoint point) {
+    std::string pName = toString(point);
+
+    if (pDecorationList->GetDecorIdByName(pName)) {
+        if (!pLevelDecorations.empty()) {
+            for (size_t i = 0; i < pLevelDecorations.size(); ++i) {
+                if (pLevelDecorations[i].uDecorationDescID == pDecorationList->GetDecorIdByName(pName)) {
+                    pParty->pos = pLevelDecorations[i].vPosition;
+                    pParty->speed = Vec3i();
+                    pParty->uFallStartZ = pParty->pos.z;
+                    pParty->_viewYaw = (TrigLUT.uIntegerHalfPi * pLevelDecorations[i].field_1A) / 90;
+                    if (pLevelDecorations[i]._yawAngle)
+                        pParty->_viewYaw = pLevelDecorations[i]._yawAngle;
+                    pParty->_viewPitch = 0;
+                }
+            }
+        }
+        if (Start_Party_Teleport_Flag) {
+            if (Party_Teleport_X_Pos)
+                pParty->pos.x = Party_Teleport_X_Pos;
+            if (Party_Teleport_Y_Pos)
+                pParty->pos.y = Party_Teleport_Y_Pos;
+            if (Party_Teleport_Z_Pos) {
+                pParty->pos.z = Party_Teleport_Z_Pos;
+                pParty->uFallStartZ = Party_Teleport_Z_Pos;
+            }
+            if (Party_Teleport_Cam_Yaw != -1)
+                pParty->_viewYaw = Party_Teleport_Cam_Yaw;
+            if (Party_Teleport_Cam_Pitch)
+                pParty->_viewPitch = Party_Teleport_Cam_Pitch;
+            if (Party_Teleport_Z_Speed)
+                pParty->speed = Vec3i(0, 0, Party_Teleport_Z_Speed);
+        }
+        Party_Teleport_Cam_Yaw = -1;
+        Start_Party_Teleport_Flag = 0;
+        Party_Teleport_Z_Speed = 0;
+        Party_Teleport_Cam_Pitch = 0;
+        Party_Teleport_Z_Pos = 0;
+        Party_Teleport_Y_Pos = 0;
+        Party_Teleport_X_Pos = 0;
+    }
 }
