+++ src/GUI/GUIWindow.cpp
@@ -24,8 +24,8 @@
 #include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/IconFrameTable.h"
 #include "Engine/Tables/AwardTable.h"
-#include "Engine/Tables/MessageScrollTable.h"
 #include "Engine/Time.h"
+#include "Engine/MapInfo.h"
 
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
