+++ src/GUI/UI/Books/CalendarBook.cpp
@@ -7,6 +7,7 @@
 #include "Engine/Localization.h"
 #include "Engine/Party.h"
 #include "Engine/AssetsManager.h"
+#include "Engine/MapInfo.h"
 
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Graphics/Viewport.h"
